@@ -6,32 +6,34 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
+  sdk: ^3.5.3
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_modular: ^6.3.2
-  cupertino_icons: ^1.0.2
+  flutter_modular: ^6.3.4
+  cupertino_icons: ^1.0.8
   flutter_speed_dial: ^7.0.0
   bordered_text: ^2.0.0
   hive: ^2.2.3  
   hive_generator: ^2.0.0
   hive_flutter: ^1.1.0
-  device_info_plus: ^9.1.0
+  device_info_plus: ^11.1.0
   rx_notifier: ^2.3.0
   badges: ^3.1.2
-  path_provider: ^2.1.1
-  dio: ^5.3.2
-  intl: ^0.18.1
-  youtube_player_iframe: ^4.0.4 
-  collection: ^1.17.2  
-  brasil_fields: ^1.14.0
+  path_provider: ^2.1.5
+  dio: ^5.7.0
+  intl: ^0.19.0
+  youtube_player_iframe: ^5.2.0 
+  collection: ^1.18.0
+  brasil_fields: ^1.15.0
+  web: ^1.1.0 
+  url_launcher_web: ^2.3.3     
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^3.0.1
+  flutter_lints: ^5.0.0
 
 flutter:
   uses-material-design: true
